Resolve "Fix compile time warnings"

Closes #207

See merge request aerys/smartshape-engine!201

+++ framework/include/minko/Signal.hpp
@@ -93,7 +93,7 @@ namespace minko
 		uint
 		numCallbacks() const
 		{
-			return _callbacks.size();
+			return static_cast<uint>(_callbacks.size());
 		}
 
 		Slot
